+++ Adapter/SQS/Consumer.php
@@ -4,6 +4,8 @@ namespace Kaliop\Queueing\Plugins\SQSBundle\Adapter\SQS;
 
 use Kaliop\QueueingBundle\Queue\MessageConsumerInterface;
 use Kaliop\QueueingBundle\Queue\ConsumerInterface;
+use Kaliop\QueueingBundle\Queue\SignalHandlingConsumerInterface;
+use Kaliop\QueueingBundle\Adapter\ForcedStopException;
 use Aws\Sqs\SqsClient;
 use Aws\TraceMiddleware;
 use Psr\Log\LoggerInterface;
@@ -11,7 +13,7 @@ use Psr\Log\LoggerInterface;
 /**
  * @todo support long polling - even though it will complicate consume() even more than it already is
  */
-class Consumer implements ConsumerInterface
+class Consumer implements ConsumerInterface, SignalHandlingConsumerInterface
 {
     /** @var  \Aws\Sqs\SqsClient */
     protected $client;
@@ -26,6 +28,10 @@ class Consumer implements ConsumerInterface
     protected $contentTypeAttribute = 'contentType';
     protected $routingAttribute = 'routingKey';
     protected $debug = false;
+    protected $forceStop = false;
+    protected $forceStopReason;
+    protected $dispatchSignals = false;
+    protected $memoryLimit = null;
 
     public function __construct(array $config)
     {
@@ -59,12 +65,13 @@ class Consumer implements ConsumerInterface
     }
 
     /**
-     * Does nothing
-     * @param int $limit
+     * @param int $limit MB
      * @return Consumer
      */
     public function setMemoryLimit($limit)
     {
+        $this->memoryLimit = $limit;
+
         return $this;
     }
 
@@ -182,6 +189,8 @@ class Consumer implements ConsumerInterface
                 }
             }
 
+            $this->maybeStopConsumer();
+
             if ($amount > 0) {
                 return;
             }
@@ -248,5 +257,37 @@ class Consumer implements ConsumerInterface
     public function getQueueUrl()
     {
         return $this->queueUrl;
+    }
+
+    public function setHandleSignals($doHandle)
+    {
+        $this->dispatchSignals = $doHandle;
+    }
+
+
+    public function forceStop($reason = '')
+    {
+        $this->forceStop = true;
+        $this->forceStopReason = $reason;
+    }
+
+    /**
+     * Dispatches signals and throws an exception if user wants to stop. To be called at execution points when there is no data loss
+     *
+     * @throws ForcedStopException
+     */
+    protected function maybeStopConsumer()
+    {
+        if ($this->dispatchSignals) {
+            pcntl_signal_dispatch();
+        }
+
+        if ($this->memoryLimit > 0 && !$this->forceStop && memory_get_usage(true) >= ($this->memoryLimit * 1024 * 1024) ) {
+            $this->forceStop("Memory limit of {$this->memoryLimit} MB reached while consuming messages");
+        }
+
+        if ($this->forceStop) {
+            throw new ForcedStopException($this->forceStopReason);
+        }
     }
 }